# apps/api/app/modules/billing/router.py
from datetime import UTC, datetime, timedelta
from decimal import Decimal
from typing import Annotated

from fastapi import APIRouter, Header, Query, Request
from sqlalchemy import func, select

from app.core.audit import set_audit_target
from app.core.db import DbSession
from app.core.errors import AppError, ErrorCode
from app.core.money import as_amount
from app.core.pagination import Page
from app.core.platform_config import get_effective_platform_config
from app.core.policies import get_effective_policies
from app.modules.account.deps import CurrentUser
from app.modules.billing import payment_service, wallet
from app.modules.billing.models import BillDailyDisk, BillHourly
from app.modules.billing.schemas import (
    BillHourlyOut,
    BillSummaryItem,
    BillSummaryOut,
    DailySummaryOut,
    LedgerEntryOut,
    PoliciesOut,
    RechargeCreate,
    RechargeOut,
    WalletOut,
)

router = APIRouter(tags=["billing"])


@router.get("/policies")
async def get_policies(session: DbSession) -> PoliciesOut:
    """计费/回收策略。公开(未登录市场页也要展示盘价);env 默认 + DB 覆盖,管理端在线调整。"""
    p = await get_effective_policies(session)
    cfg = await get_effective_platform_config(session)
    return PoliciesOut(
        disk_price_gb_month=p.disk_price_gb_month,
        disk_min_gb=p.disk_min_gb,
        disk_max_gb=p.disk_max_gb,
        disk_grace_days=p.disk_grace_days,
        disk_frozen_days=p.disk_frozen_days,
        freeze_grace_hours=p.freeze_grace_hours,
        low_balance_warn_hours_default=p.low_balance_warn_hours,
        real_name_required_for_recharge=cfg["real_name_required_for_recharge"] == "true",
    )


# 本地日界的偏移量(分)。与 /bills/daily-summary 同款:默认 480 = 东八区,
# 前端一律传浏览器真实 offset。
TzOffset = Query(default=480, ge=-720, le=840)


def _parse_month(month: str, tz_offset_minutes: int) -> tuple[datetime, datetime]:
    """月窗口按本地月初/次月初切,与日账单同口径(按 UTC 切会让日账单加总 ≠ 月账单)。"""
    try:
        local_start = datetime.strptime(month, "%Y-%m").replace(tzinfo=UTC)
    except ValueError as exc:
        raise AppError(ErrorCode.VALIDATION_ERROR, key="billing.badMonthFormat") from exc
    local_end = (
        local_start.replace(year=local_start.year + 1, month=1)
        if local_start.month == 12
        else local_start.replace(month=local_start.month + 1)
    )
    offset = timedelta(minutes=tz_offset_minutes)
    return local_start - offset, local_end - offset


@router.get("/wallet")
async def get_wallet(user: CurrentUser, session: DbSession) -> WalletOut:
    w = await wallet.get_or_create_wallet(session, user.id)
    await session.commit()
    return WalletOut.model_validate(w)


@router.get("/wallet/ledger")
async def get_ledger(
    user: CurrentUser,
    session: DbSession,
    cursor: str | None = None,
    limit: int | None = Query(default=None, le=100),
) -> Page[LedgerEntryOut]:
    return await wallet.ledger_page(session, user.id, cursor=cursor, limit=limit)


@router.get("/bills/hourly")
async def list_hourly_bills(
    user: CurrentUser,
    session: DbSession,
    instance_id: int | None = None,
    month: str | None = None,
    tz_offset_minutes: int = TzOffset,
    cursor: str | None = None,
    limit: int | None = Query(default=None, le=100),
) -> Page[BillHourlyOut]:
    return await wallet.hourly_bills_page(
        session,
        user.id,
        instance_id=instance_id,
        month_range=_parse_month(month, tz_offset_minutes) if month is not None else None,
        cursor=cursor,
        limit=limit,
    )


@router.get("/bills/summary")
async def bill_summary(
    user: CurrentUser, session: DbSession, month: str, tz_offset_minutes: int = TzOffset
) -> BillSummaryOut:
    """月度汇总 + 按实例成本归因(消费概览环图数据源)。窗口按本地月界切。"""
    start, end = _parse_month(month, tz_offset_minutes)
    gpu_rows = (
        (
            await session.execute(
                select(
                    BillHourly.instance_id,
                    func.sum(BillHourly.amount),
                    func.sum(BillHourly.seconds_used),
                )
                .where(
                    BillHourly.user_id == user.id,
                    BillHourly.hour_start >= start,
                    BillHourly.hour_start < end,
                )
                .group_by(BillHourly.instance_id)
            )
        )
        .tuples()
        .all()
    )
    disk_total = (
        await session.execute(
            select(func.coalesce(func.sum(BillDailyDisk.amount), 0)).where(
                BillDailyDisk.user_id == user.id,
                BillDailyDisk.day >= start,
                BillDailyDisk.day < end,
            )
        )
    ).scalar_one()
    items = [
        BillSummaryItem(
            instance_id=iid, total_amount=amount or Decimal("0.00"), total_seconds=int(secs or 0)
        )
        for iid, amount, secs in gpu_rows
    ]
    gpu_total = sum((i.total_amount for i in items), Decimal("0.00"))
    return BillSummaryOut(
        month=month, gpu_total=gpu_total, disk_total=Decimal(disk_total), items=items
    )


@router.get("/bills/daily-summary")
async def bill_daily_summary(
    user: CurrentUser,
    session: DbSession,
    date: str,
    tz_offset_minutes: int = Query(default=480, ge=-720, le=840),
) -> DailySummaryOut:
    """当日消费(实例列表「今日 ¥Y.YY」与费用中心数据源),本地日界经 tz_offset 折算。"""
    # hour_start 为 UTC 整点,offset 为整分时窗口边界不会切开小时账单
    try:
        local_midnight = datetime.strptime(date, "%Y-%m-%d").replace(tzinfo=UTC)
    except ValueError as exc:
        raise AppError(ErrorCode.VALIDATION_ERROR, key="billing.badDateFormat") from exc
    start = local_midnight - timedelta(minutes=tz_offset_minutes)
    end = start + timedelta(days=1)
    gpu_rows = (
        (
            await session.execute(
                select(
                    BillHourly.instance_id,
                    func.sum(BillHourly.amount),
                    func.sum(BillHourly.seconds_used),
                )
                .where(
                    BillHourly.user_id == user.id,
                    BillHourly.hour_start >= start,
                    BillHourly.hour_start < end,
                )
                .group_by(BillHourly.instance_id)
            )
        )
        .tuples()
        .all()
    )
    disk_total = (
        await session.execute(
            select(func.coalesce(func.sum(BillDailyDisk.amount), 0)).where(
                BillDailyDisk.user_id == user.id,
                BillDailyDisk.day >= start,
                BillDailyDisk.day < end,
            )
        )
    ).scalar_one()
    items = [
        BillSummaryItem(
            instance_id=iid, total_amount=amount or Decimal("0.00"), total_seconds=int(secs or 0)
        )
        for iid, amount, secs in gpu_rows
    ]
    gpu_total = sum((i.total_amount for i in items), Decimal("0.00"))
    return DailySummaryOut(
        date=date, gpu_total=gpu_total, disk_total=as_amount(Decimal(disk_total)), items=items
    )


# ---------- 充值 ----------


@router.post("/wallet/recharges", status_code=201)
async def create_recharge(
    body: RechargeCreate,
    user: CurrentUser,
    session: DbSession,
    request: Request,
    idempotency_key: Annotated[str | None, Header(alias="Idempotency-Key")] = None,
) -> RechargeOut:
    cfg = await get_effective_platform_config(session)
    if cfg["real_name_required_for_recharge"] == "true" and user.verification_status != "verified":
        raise AppError(
            ErrorCode.REAL_NAME_REQUIRED,
            key="billing.realNameRequiredForRecharge",
            http_status=403,
        )
    order = await payment_service.create_recharge(
        session, user.id, body.amount, body.channel, idempotency_key
    )
    set_audit_target(request, f"order:{order.order_no}")
    return RechargeOut.model_validate(order)


@router.get("/wallet/recharges/{order_no}")
async def get_recharge(order_no: str, user: CurrentUser, session: DbSession) -> RechargeOut:
    order = await payment_service.get_order(session, user.id, order_no)
    return RechargeOut.model_validate(order)