# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "BillHourlyOut": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "gpu_count": {
            "title": "Gpu Count",
            "type": "integer"
          },
          "hour_start": {
            "format": "date-time",
            "title": "Hour Start",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "instance_id": {
            "title": "Instance Id",
            "type": "integer"
          },
          "seconds_used": {
            "title": "Seconds Used",
            "type": "integer"
          },
          "unit_price": {
            "title": "Unit Price",
            "type": "string"
          }
        },
        "required": [
          "id",
          "instance_id",
          "hour_start",
          "seconds_used",
          "unit_price",
          "gpu_count",
          "amount"
        ],
        "title": "BillHourlyOut",
        "type": "object"
      },
      "BillSummaryItem": {
        "properties": {
          "instance_id": {
            "title": "Instance Id",
            "type": "integer"
          },
          "instance_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance Name"
          },
          "total_amount": {
            "title": "Total Amount",
            "type": "string"
          },
          "total_seconds": {
            "title": "Total Seconds",
            "type": "integer"
          }
        },
        "required": [
          "instance_id",
          "total_amount",
          "total_seconds"
        ],
        "title": "BillSummaryItem",
        "type": "object"
      },
      "BillSummaryOut": {
        "properties": {
          "disk_total": {
            "title": "Disk Total",
            "type": "string"
          },
          "gpu_total": {
            "title": "Gpu Total",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillSummaryItem"
            },
            "title": "Items",
            "type": "array"
          },
          "month": {
            "title": "Month",
            "type": "string"
          }
        },
        "required": [
          "month",
          "gpu_total",
          "disk_total",
          "items"
        ],
        "title": "BillSummaryOut",
        "type": "object"
      },
...
      "DailySummaryOut": {
        "description": "当日消费汇总(本地日界由 tz_offset_minutes 折算)。",
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "disk_total": {
            "title": "Disk Total",
            "type": "string"
          },
          "gpu_total": {
            "title": "Gpu Total",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillSummaryItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "date",
          "gpu_total",
          "disk_total",
          "items"
        ],
        "title": "DailySummaryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LedgerEntryOut": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "balance_after": {
            "title": "Balance After",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ref_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Id"
          },
          "ref_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Type"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "amount",
          "balance_after",
          "ref_type",
          "ref_id",
          "remark",
          "created_at"
        ],
        "title": "LedgerEntryOut",
        "type": "object"
      },
...
      "Page_BillHourlyOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillHourlyOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[BillHourlyOut]",
        "type": "object"
      },
      "Page_LedgerEntryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[LedgerEntryOut]",
        "type": "object"
      },
...
      "RechargeCreate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "channel": {
            "default": "mock",
            "title": "Channel",
            "type": "string"
          }
        },
        "required": [
          "amount"
        ],
        "title": "RechargeCreate",
        "type": "object"
      },
      "RechargeOut": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "order_no": {
            "title": "Order No",
            "type": "string"
          },
          "qr_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Url"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "order_no",
          "amount",
          "channel",
          "status",
          "qr_url",
          "expires_at",
          "created_at"
        ],
        "title": "RechargeOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WalletOut": {
        "properties": {
          "balance": {
            "title": "Balance",
            "type": "string"
          },
          "frozen_amount": {
            "title": "Frozen Amount",
            "type": "string"
          }
        },
        "required": [
          "balance",
          "frozen_amount"
        ],
        "title": "WalletOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SuperDL API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bills/daily-summary": {
      "get": {
        "description": "当日消费(实例列表「今日 ¥Y.YY」与费用中心数据源),本地日界经 tz_offset 折算。",
        "operationId": "bill_daily_summary_api_v1_bills_daily_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tz_offset_minutes",
            "required": false,
            "schema": {
              "default": 480,
              "maximum": 840,
              "minimum": -720,
              "title": "Tz Offset Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bill Daily Summary",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/bills/hourly": {
      "get": {
        "operationId": "list_hourly_bills_api_v1_bills_hourly_get",
        "parameters": [
          {
            "in": "query",
            "name": "instance_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instance Id"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "in": "query",
            "name": "tz_offset_minutes",
            "required": false,
            "schema": {
              "default": 480,
              "maximum": 840,
              "minimum": -720,
              "title": "Tz Offset Minutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BillHourlyOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Hourly Bills",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/bills/summary": {
      "get": {
        "description": "月度汇总 + 按实例成本归因(消费概览环图数据源)。窗口按本地月界切。",
        "operationId": "bill_summary_api_v1_bills_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "title": "Month",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tz_offset_minutes",
            "required": false,
            "schema": {
              "default": 480,
              "maximum": 840,
              "minimum": -720,
              "title": "Tz Offset Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bill Summary",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/api/v1/wallet": {
      "get": {
        "operationId": "get_wallet_api_v1_wallet_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Wallet",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/wallet/ledger": {
      "get": {
        "operationId": "get_ledger_api_v1_wallet_ledger_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_LedgerEntryOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Ledger",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/wallet/recharges": {
      "post": {
        "operationId": "create_recharge_api_v1_wallet_recharges_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RechargeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RechargeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Recharge",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/wallet/recharges/{order_no}": {
      "get": {
        "operationId": "get_recharge_api_v1_wallet_recharges__order_no__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_no",
            "required": true,
            "schema": {
              "title": "Order No",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RechargeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Recharge",
        "tags": [
          "billing"
        ]
      }
    },
...
  }
}